requirements.txt
pip
fastapi
pandas
numpy
apache-airflow-providers-common-sql
apache-airflow-providers-postgres
apache-airflow-providers-mysql
apache-airflow-providers-google
apache-airflow-providers-common-io
apache-airflow[postgres]
